Create README.md
Collinearity among environmental and ecological predictors is widespread in ecology, yet posses serious issues in statistical analyses. Complete subsets information theoretic approaches are becoming an increasingly popular tool for exploring predictive power and variable importance where a wide range of candidate predictors are being considered.
We developed a simple function in the statistical programming language R that can be used to construct, fit and compare a complete model set of possible ecological or environmental predictors, given a response variable of interest. The function is based on Generalized Additive Models (GAM) and builds on the MuMIn package. 
Advantages include the capacity to fit more predictors than there are replicates, automatic removal of collinear models, and model sets that include interactions between factors and smooth predictors. The function takes a range of arguments that allow reasonable control over the model set being constructed, including specifying cyclic and linear continuous predictors, specification of the smoothing algorithm used and the maximum complexity allowed for smooth terms. 
The use of the function is demonstrated via three case studies that serve to highlight the broader utility of the approach for exploratory ecology.